Initial commit of C++ code for the MWN-OU diffusion
C++ functions for the log-likelihood of a set of pairs and for the
evaluation and sampling of the stationary density. LogLikWNOUPairs.cpp,
DensStatWNOU.cpp and SampStatWNOU.cpp are the subroutines for each
task. maintpd.cpp, maindens.cpp and mainsamo.cpp are the callable
programs from console (see readme.md). system.call.R provides a small
example of how to perform such system calls from R.